Re-organize 'FileHashStore' test methods into correct test classes and update exception handling in 'FileHashStore'
